requirements.txt
APScheduler==3.9.1.post1
blinker==1.5
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
Werkzeug==2.2.2